Re-work TaskManager
TaskManager will now use ThreadPoolExecutor to manage threads.
After doing some reading, this is the preferred method of handling
multiple threads in python. The logic for the save methods was also
changed to include the use of Rlock to isolate "critical" sections of
the code (i.e. pieces that need to be run sequentially)